% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnng.R
\name{gnng}
\alias{gnng}
\title{The group LARS method to select factors in liner regression problem raised by Yuan and Lin(2004)}
\usage{
gnng(X, Y, sigma = -1, group, scale = FALSE)
}
\arguments{
\item{X}{the design matrix. Some details about X are shown in group}

\item{Y}{the response variable}

\item{sigma}{the standard deviation of the noise term to calculate Cp statistic of selected models. If sigma = -1 the OLS estimate deviation will be used in the progress. Initial -1}

\item{group}{a numeric vector of the quantity of variables in each factor(group). Thus the variables in one factor are asked to be arranged contiguously in the design matrix X}

\item{scale}{logical. If TRUE, each variable is standardized to have unit L2 norm, otherwise it is left alone.}
}
\value{
a list abput the solution path, including the selectived coefficient, the Cp statistic of each model and the time the method cost.
}
\description{
The group non-negetive garrotte method to select factors in liner regression problem raised by Yuan and Lin(2004)
}
\examples{
\dontrun{
  Y = coef \%*\% X + rnorm(n, 0, 0.5)
  spath = gnng(X, Y, sigma = 0.5, group = rep(2, 15))
 }
}
