\name{StatComp22005-package}
\alias{StatComp22005-package}
\docType{package}
\title{Final homework for 'Statistical Computing' course.}
\description{
Final homework for 'Statistical Computing' course. It also includes two R functions to solving factor select problem in linear model and a Cpp function comes from my homework which generate a gibbs sample.
}
\author{
Tianyi Zhang
Maintainer:Tianyi Zhang <3320676831@qq.com>
}
\keyword{package}
